Move under 'pytest_catchlog' package, extract fixture.py and common.py
