+++ pytest_catchlog/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+__version__ = '1.2.2'
